+++ src/composables/dom/useRefreshToken.ts
@@ -1,5 +1,5 @@
 import { computed } from 'vue'
-import { useRefreshTokens } from '/@/store/app/refreshTokens'
+import { useRefreshTokens } from '/@/store/ui/refreshTokens'
 
 export const useRefreshToken = (key: string) => {
   const { refresh: refreshImpl, getToken } = useRefreshTokens()
